# docs/swagger.yaml
basePath: /
definitions:
...
  GetSummaryByEANRow:
    properties:
      active_ingredient:
        type: string
      adverse_effects:
        type: string
      brand_name:
        type: string
      contraindications:
        type: string
      description:
        type: string
      drug_interactions:
        type: string
      ean:
        type: string
      leaflet_expedient:
        type: string
      leaflet_published_at:
        type: string
      manufacturer:
        type: string
      mechanism_of_action:
        type: string
      missed_dose:
        type: string
      posology:
        type: string
      presentation_registration:
        type: string
      registration_number:
        type: string
      side_effects:
        type: string
      source_url:
        type: string
      storage:
        type: string
      warnings:
        type: string
      what_is_it_for:
        type: string
      when_to_seek_help:
        type: string
    type: object
...
  ListDrugsRow:
    properties:
      active_ingredient:
        type: string
      brand_name:
        type: string
      id:
        type: integer
      manufacturer:
        type: string
      registration_number:
        type: string
      updated_at:
        type: string
    type: object
...
  Role:
    enum:
    - editor
    - reviewer
    - admin
    type: string
    x-enum-comments:
      RoleAdmin: + manages users
      RoleEditor: creates and edits drugs, packages and leaflets
      RoleReviewer: + marks leaflets as reviewed (pharmacist)
    x-enum-descriptions:
    - creates and edits drugs, packages and leaflets
    - + marks leaflets as reviewed (pharmacist)
    - + manages users
    x-enum-varnames:
    - RoleEditor
    - RoleReviewer
    - RoleAdmin
  User:
    properties:
      email:
        example: ana@farmacia.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: Ana Souza
        type: string
      role:
        allOf:
        - $ref: '#/definitions/Role'
        example: reviewer
    type: object
  changePasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    type: object
  createDrugRequest:
    properties:
      active_ingredient:
        type: string
      brand_name:
        type: string
      manufacturer:
        type: string
      registration_number:
        type: string
    type: object
...
  idResponse:
    properties:
      id:
        example: 1
        type: integer
    type: object
  listResponse:
    properties:
      data: {}
      limit:
        type: integer
      offset:
        type: integer
    type: object
  loginRequest:
    properties:
      email:
        example: ana@farmacia.com
        type: string
      password:
        example: uma-senha-forte
        type: string
    type: object
...
info:
  contact: {}
  description: |-
    API that reads the barcode on a medicine box and returns the simplified leaflet.
    Public route: GET /drugs/ean/{ean}. All others require login (POST /auth/login sets a session cookie).
  title: PharmaCode API
  version: "1.0"
paths:
  /auth/login:
    post:
      consumes:
      - application/json
      description: Checks email and password and sets the session cookie (HttpOnly).
        In Swagger, after logging in the other routes work in this same browser.
      parameters:
      - description: Credentials
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/loginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
        "400":
          description: invalid json
          schema:
            type: string
        "401":
          description: invalid email or password
          schema:
            type: string
        "429":
          description: too many attempts
          schema:
            type: string
      summary: Log in to the admin panel
      tags:
      - auth
...
  /auth/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/User'
        "401":
          description: authentication required
          schema:
            type: string
      summary: Current user
      tags:
      - auth
  /auth/password:
    put:
      consumes:
      - application/json
      description: Requires the current password. The user's other sessions are ended;
        the current one stays.
      parameters:
      - description: Current and new password (8 to 72 characters)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/changePasswordRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: weak new password / invalid json
          schema:
            type: string
        "401":
          description: authentication required
          schema:
            type: string
        "403":
          description: current password is wrong
          schema:
            type: string
      summary: Change your own password
      tags:
      - auth
  /drugs:
    get:
      parameters:
      - description: Items per page (default 20, max 100)
        in: query
        name: limit
        type: integer
      - description: How many items to skip (default 0)
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/listResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/ListDrugsRow'
                  type: array
              type: object
        "400":
          description: invalid limit / invalid offset
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: List drugs (paginated)
      tags:
      - drugs
    post:
      consumes:
      - application/json
      parameters:
      - description: Drug data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/createDrugRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/idResponse'
        "400":
          description: invalid json or missing required field
          schema:
            type: string
        "409":
          description: registration_number already exists
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Create a drug
      tags:
      - drugs
  /drugs/{id}:
    delete:
      parameters:
      - description: Drug ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "400":
          description: invalid id
          schema:
            type: string
        "404":
          description: drug not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Delete a drug
      tags:
      - drugs
    put:
      consumes:
      - application/json
      parameters:
      - description: Drug ID
        in: path
        name: id
        required: true
        type: integer
      - description: Drug data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/createDrugRequest'
      responses:
        "204":
          description: No Content
        "400":
          description: invalid id or json
          schema:
            type: string
        "404":
          description: drug not found
          schema:
            type: string
        "409":
          description: registration_number already exists
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Update a drug
      tags:
      - drugs
  /drugs/ean/{ean}:
    get:
      description: Returns the drug data and the leaflet summary from the barcode
        on the box
      parameters:
      - description: Package EAN barcode
        in: path
        name: ean
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/GetSummaryByEANRow'
        "400":
          description: ean is required
          schema:
            type: string
        "404":
          description: ean not found
          schema:
            type: string
        "500":
          description: internal server error
          schema:
            type: string
      summary: Get the simplified leaflet by EAN
      tags:
      - drugs
...
swagger: "2.0"